requirements.txt
fastapi>=0.109.0
httpx>=0.26.0
uvicorn[standard]>=0.27.0
pandas>=2.0.0
scikit-learn>=1.3.0
stripe>=7.0.0
python-multipart>=0.0.6
supabase>=2.0.0
python-dotenv>=1.0.0
